Create actor_json.py and update JSON schema
- Move JSON handling into separate file
- Separate "actions" and "queries" in JSON schema
