+++ eventeditor/__main__.py
@@ -9,9 +9,9 @@ import typing
 import evfl
 from evfl import EventFlow
 import eventeditor.ai as ai
+import eventeditor.actor_json as aj
 from eventeditor.actor_view import ActorView
 from eventeditor.event_view import EventView
-import eventeditor.event_edit_dialog as evedit
 from eventeditor.flow_data import FlowData, FlowDataChangeReason
 from eventeditor.flowchart_view import FlowchartView
 import eventeditor.util as util
@@ -181,7 +181,7 @@ class MainWindow(q.QMainWindow):
     def readSettings(self) -> None:
         settings = qc.QSettings()
         ai.set_rom_path(settings.value('paths/rom_root'))
-        evedit.set_actor_json_path(settings.value('paths/actor_json_root'))
+        aj.set_actor_json_path(settings.value('paths/actor_json_root'))
         settings.beginGroup('MainWindow')
         self.resize(settings.value('size', qc.QSize(800, 600)))
         self.move(settings.value('pos', qc.QPoint(200, 200)))
